skills/test/ai-intf-test/requirements.txt
openpyxl
PyYAML
requests
